valuation-agent/requirements.txt
Flask==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.1
pydantic>=2.8.0
requests>=2.32.0

langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-anthropic>=0.3.0
langchain-groq==0.3.0
langgraph>=0.2.0
tavily-python==0.7.8

rapidfuzz>=3.0.0
numpy>=1.24.0
